LayoutTests/http/tests/security/contentSecurityPolicy/object-src-no-url-blocked-expected.txt

CONSOLE MESSAGE: Refused to load because it does not appear in the object-src directive of the Content Security Policy.
This test passes if there is a console message saying the plugin was blocked.
